Move ProtocolVersion and Capabilities to reth-eth-wire-types
