+++ crates/net/eth-wire-types/src/lib.rs
@@ -12,7 +12,7 @@ mod status;
 pub use status::{Status, StatusBuilder};
 
 pub mod version;
-pub use version::EthVersion;
+pub use version::{EthVersion, ProtocolVersion};
 
 pub mod message;
 pub use message::{EthMessage, EthMessageID, ProtocolMessage};
@@ -37,3 +37,6 @@ pub use receipts::*;
 
 pub mod disconnect_reason;
 pub use disconnect_reason::*;
+
+pub mod capability;
+pub use capability::*;
