+++ crates/net/eth-wire/src/p2pstream.rs
@@ -16,7 +16,7 @@ use reth_primitives::{
 };
 use std::{
     collections::VecDeque,
-    fmt, io,
+    io,
     pin::Pin,
     task::{ready, Context, Poll},
     time::Duration,
@@ -788,49 +788,10 @@ impl TryFrom<u8> for P2PMessageID {
     }
 }
 
-/// RLPx `p2p` protocol version
-#[derive_arbitrary(rlp)]
-#[derive(Copy, Clone, Debug, Default, PartialEq, Eq)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-pub enum ProtocolVersion {
-    /// `p2p` version 4
-    V4 = 4,
-    /// `p2p` version 5
-    #[default]
-    V5 = 5,
-}
-
-impl fmt::Display for ProtocolVersion {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "v{}", *self as u8)
-    }
-}
-
-impl Encodable for ProtocolVersion {
-    fn encode(&self, out: &mut dyn BufMut) {
-        (*self as u8).encode(out)
-    }
-    fn length(&self) -> usize {
-        // the version should be a single byte
-        (*self as u8).length()
-    }
-}
-
-impl Decodable for ProtocolVersion {
-    fn decode(buf: &mut &[u8]) -> alloy_rlp::Result<Self> {
-        let version = u8::decode(buf)?;
-        match version {
-            4 => Ok(Self::V4),
-            5 => Ok(Self::V5),
-            _ => Err(RlpError::Custom("unknown p2p protocol version")),
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{capability::SharedCapability, test_utils::eth_hello, EthVersion};
+    use crate::{capability::SharedCapability, test_utils::eth_hello, EthVersion, ProtocolVersion};
     use tokio::net::{TcpListener, TcpStream};
     use tokio_util::codec::Decoder;
